preliminary implementation of kernel selection.
