+++ src/rocprof_compute_tui/widgets/center_panel/center_area.py
@@ -29,9 +29,10 @@ Contains the panel widgets used in the main layout.
 """
 
 from textual.containers import Vertical
-from textual.widgets import Label, TabPane
+from textual.widgets import TabPane
 
-from rocprof_compute_tui.widgets.center_panel.analyze_view import AnalyzeView
+from rocprof_compute_tui.views.analyze_view import AnalyzeView
+from rocprof_compute_tui.views.kernel_view import KernelView
 from rocprof_compute_tui.widgets.tabbed_content import TabsTabbedContent
 
 
@@ -49,14 +50,14 @@ class CenterPanel(Vertical):
 
         self.default_tab = "center-analyze"
         self.analyze_view = AnalyzeView()
+        self.kernel_view = KernelView()
 
     def compose(self):
         with TabsTabbedContent(initial="tab-analyze"):
             with TabPane("Basic View", id="tab-analyze"):
                 yield self.analyze_view
-            # TODO:
-            # with TabPane("placeholder (🚧)", id="tab-1"):
-            #     yield Label("🚧 Under Construction")
+            with TabPane("Kernel View", id="tab-kernel"):
+                yield self.kernel_view
 
     def on_mount(self) -> None:
         self.add_class("section")
